% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esample_data.R
\docType{data}
\name{BRCA_metadata}
\alias{BRCA_metadata}
\title{TCGA Breast Invasive Carcinoma sample data}
\format{
A data frame with 105 rows and 1 variable (IDs are in rownames):
\describe{
\item{SUBTYPE}{The breast cancer subtype}
}
}
\usage{
BRCA_metadata
}
\description{
A dataset containing sample data for 105 breast tissue biopsies from the
TCGA-BRCA project
}
\keyword{datasets}
