% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esample_data.R
\docType{data}
\name{BRCA_normCounts}
\alias{BRCA_normCounts}
\title{TCGA Breast Invasive Carcinoma (BRCA) normalised gene expression data}
\format{
A data frame with 19737 rows representing normalised gene counts
and 105 columns representing samples.
}
\usage{
BRCA_normCounts
}
\description{
A dataset containing the z scored gene expression data for 105 breast
biopsies from the TCGA-BRCA project
}
\keyword{datasets}
