% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_core_functions.R
\name{calc_pvalues_network}
\alias{calc_pvalues_network}
\title{Calculate the pvalues for specific subgroup network samples}
\usage{
calc_pvalues_network(
  normalised_counts,
  metadata,
  sig_var,
  subgroup_variable,
  subgroups_length,
  regression_method = "rlm",
  subgroup_network
)
}
\arguments{
\item{normalised_counts}{a data frame containing normalised counts from an
high throughput sequencing experiment.
Sample IDs must be in columns and gene/miRNA/TFs in rows.
Objects of class \code{matrix} are not allowed.}

\item{metadata}{a data frame of sample data with rownames matching the
sample IDs in \code{normalised_counts} colnames}

\item{sig_var}{Inherited from \code{generate_subnetworks}. It can be
\code{q.value} or \code{p.value} depending on how \code{use_qvalues} was set in the
\code{generate_subnetworks} function (default \code{FALSE}).}

\item{subgroup_variable}{column name in \code{metadata} that contains the
subgroup identifier for each sample in \code{normalised_counts}}

\item{subgroups_length}{integer number indicating the number of subgroups}

\item{regression_method}{whether to use robust linear modelling to obtain
p value of the interactions. Options are 'rlm' (default) or 'lm'}

\item{subgroup_network}{network table for a specific subgroup}
}
\value{
a list of p values
}
\description{
Calculate the pvalues for specific subgroup network samples
}
