% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_core_functions.R
\docType{class}
\name{deggs-class}
\alias{deggs-class}
\title{An S4 class to define the deggs output}
\description{
An S4 class to define the deggs output
}
\section{Slots}{

\describe{
\item{\code{subnetworks}}{a list of data frames containing the subgroup networks and
a number indicating the total count of links that reached significance in all
subgroups (considering p or adj p < 0.05)}

\item{\code{normalised_counts}}{a data frame containig the normalised count data
given in input.}

\item{\code{metadata}}{a data frame of sample data given in input and matching
the sample IDs in \code{normalised_counts}.}

\item{\code{subgroup_variable}}{column name in \code{metadata} that contains the
subgroup definition for each sample in \code{normalised_counts}.}

\item{\code{regression_method}}{the chosen regression method used to calculate
interaction p values. It can be either 'rlm' or 'lm'.}

\item{\code{subgroups}}{character vector indicating which subgroups
are used for comparison.}

\item{\code{use_qvalues}}{logical. Indicates whether p values are adjusted via
Storey's q value method}
}}

