% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_core_functions.R
\name{generate_subnetworks}
\alias{generate_subnetworks}
\title{Generate subnetworks}
\usage{
generate_subnetworks(
  normalised_counts,
  metadata,
  subgroup_variable,
  regression_method = "rlm",
  subgroups = NULL,
  entrezIDs = FALSE,
  convert_to_gene_symbols = TRUE,
  use_qvalues = TRUE,
  cores = parallel::detectCores()/2
)
}
\arguments{
\item{normalised_counts}{a data frame containig the normalised sequencing
count data. Sample IDs must be in columns and gene/miRNA/TFs in rows.
Objects of class \code{matrix} are not allowed.}

\item{metadata}{a data frame of sample information matching the sample IDs in
\code{normalised_counts}}

\item{subgroup_variable}{column name in \code{metadata} which contains the
subgroup definition for each sample in \code{normalised_counts}}

\item{regression_method}{whether to use robust linear modeling for the
interaction p-values. Options are 'rlm' (default) or 'lm'}

\item{subgroups}{optional character vector indicating which subgroups
are used for comparison. If not specified, all subgroups in
\code{subgroup_variable} will be considered}

\item{entrezIDs}{Logical whether gene ids in \code{normalised_counts} are in
entrez id (TRUE) or gene symbols (FALSE). Default FALSE}

\item{convert_to_gene_symbols}{Logical to be used when using entrez ids.
If TRUE (default), the output will show gene symbols}

\item{use_qvalues}{Whether to use Storey's q-values. If FALSE, unadjusted
p-values will be used.}

\item{cores}{Number of cores to use.}
}
\value{
a Deggs object with subgroup specific networks incorporating pvalues
for each interaction
}
\description{
Generate subgroup specific gene-gene interaction networks with interaction
pvalues
}
