% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{node_boxplot}
\alias{node_boxplot}
\title{Boxplots of single nodes across subgroups}
\usage{
node_boxplot(gene, deggs_object)
}
\arguments{
\item{gene}{gene name  (must be contained in rownames(normalised_counts))}

\item{cex.lab}{lab dimension}
}
\value{
the gene boxplot
}
\description{
This function is used only by \code{View_interactive_subnetwork}
}
