% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{print_simple_network}
\alias{print_simple_network}
\title{View subnetworks in basic igraph style}
\usage{
print_simple_network(deggs_object, subgroup)
}
\arguments{
\item{deggs_object}{an object of class Deggs generated from
\code{generate_subnetworks}}

\item{subgroup}{the subgroup of interest. The associated network will be
visualised}
}
\value{
an igraph plot of the selected subnetwork
}
\description{
View igraph style networks for a specific subgroup
}
