% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esample_data.R
\docType{data}
\name{subnetworks_object}
\alias{subnetworks_object}
\title{A \code{generate_subnetworks} sample output}
\format{
A data frame with 105 rows and 1 variable (IDs are in rownames):
\describe{
\item{subnetworks}{List of subgroup networks + number of total links
reaching significance (< 0.05) in their associated interaction p-value. }
\item{metadata}{The input sample data.}
\item{normalised_counts}{The input expression data.}
}
}
\usage{
subnetworks_object
}
\description{
A DEGGs object with subgroup specific networks incorporating p-values
for each interaction
}
\keyword{datasets}
