% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_interactive_subnetwork.R
\name{View_interactive_subnetwork}
\alias{View_interactive_subnetwork}
\title{View interactive shiny app style subnetworks}
\usage{
View_interactive_subnetwork(DEGGs_object, subgroup, subgroup_variable)
}
\arguments{
\item{DEGGs_object}{an object of class DEGGs generated from
\code{generate_subnetworks}}

\item{subgroup}{the subgroup of interest: the associated network will be
visualised}
}
\value{
a network plot with selectable nodes and links
}
\description{
Explore subnetworks and interactively select regression plots
}
