% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_functions.R
\name{calc_pvalues_network}
\alias{calc_pvalues_network}
\title{Calculate the pvalues for specific subgroup network samples}
\usage{
calc_pvalues_network(
  subgroup_network,
  normalised_counts,
  metadata,
  subgroup_variable,
  regression_method = "rlm",
  subgroups_length
)
}
\arguments{
\item{subgroup_network}{A igraph network related to a specific subgroup samples}

\item{normalised_counts}{A dataframe that represent the gene expressions
of samples}

\item{metadata}{A dataframe that represent the subgroup of samples}

\item{subgroup_variable}{A string that represent the column name of the subgroup
inside metadata the dataframe}

\item{regression_method}{whether to use robust linear modeling for the
interaction p-values. Options are 'rlm' (default) or 'lm'}

\item{subgroups_length}{An integer number that represent the number
of subgroups inside the metadata dataframe}
}
\value{
The list of pvalues
}
\description{
Calculate the pvalues for specific subgroup network samples
}
