% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_functions.R
\name{calc_pvalues_percentile}
\alias{calc_pvalues_percentile}
\title{Compute interaction pvalues for a percentile}
\usage{
calc_pvalues_percentile(
  normalised_counts,
  metadata,
  percentile,
  subgroups_df_list,
  permutations,
  regression_method = "rlm",
  edges,
  subgroup_variable,
  subgroups_length
)
}
\arguments{
\item{percentile}{A float number that represents the percentile}

\item{subgroups_df_list}{list of subgroup dataframes}

\item{permutations}{dataframe containing the subgroups permutations in rows}

\item{regression_method}{whether to use robust linear modeling for the
interaction p-values. Options are 'rlm' (default) or 'lm'}

\item{edges}{A dataframe of the meta pathway edges}

\item{subgroup_variable}{A string that represent the column name of the subgroup
inside metadata the dataframe}

\item{subgroups_length}{An integer number that represent the number
of subgroups inside the metadata dataframe}
}
\value{
The list of float numbers of the significant pvalues
for a specific percentile
}
\description{
Compute interaction pvalues for a percentile
}
