% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_core_functions.R
\docType{class}
\name{Deggs-class}
\alias{Deggs-class}
\title{An S4 class to define the Deggs output}
\description{
An S4 class to define the Deggs output
}
\section{Slots}{

\describe{
\item{\code{subnetworks}}{a list of specific networks in the form of tables (data.frame),
and the count of significant interactions}

\item{\code{normalised_counts}}{a data frame containig the normalised tidied
count data.}

\item{\code{metadata}}{a tidied data frame of sample information matching sample IDs
in \code{normalised_counts}.}

\item{\code{subgroup_variable}}{column name in \code{metadata} which contains the
subgroup definition for each sample in \code{normalised_counts}.}

\item{\code{regression_method}}{the chosen regression method used to calculate
interaction p values. It can be either 'rlm' or 'lm'.}

\item{\code{subgroups}}{character vector indicating which subgroups
are used for comparison.}
}}

