% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{node_boxplot}
\alias{node_boxplot}
\title{Boxplots of expression levels of single nodes (genes) per subgroups}
\usage{
node_boxplot(gene, deggs_object)
}
\arguments{
\item{gene}{gene name  (must be contained in \code{rownames(normalised_counts)})}

\item{deggs_object}{an object of class \code{deggs} generated from
\code{generate_subnetworks}}
}
\value{
the gene boxplot
}
\description{
This function is used only by \code{View_interactive_subnetwork}
}
