% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_functions.R
\name{print_force_network}
\alias{print_force_network}
\title{View subnetworks in network3D style}
\usage{
print_force_network(DEGGs_object, subgroup)
}
\arguments{
\item{DEGGs_object}{an object of class DEGGs generated from
\code{generate_subnetworks}}

\item{subgroup}{the subgroup of interest. The associated network will be
visualised}
}
\value{
a network3D plot of the selected subnetwork
}
\description{
View network3D style networks for a specific subgroup
}
