% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{print_regressions}
\alias{print_regressions}
\title{Print the regression graph of a specific genes}
\usage{
print_regressions(
  deggs_object,
  gene_A,
  gene_B,
  use_qvalues = TRUE,
  legend_position = "topright"
)
}
\arguments{
\item{gene_A}{A string of gene selected}

\item{gene_B}{A string of gene selected}

\item{legend_position}{posion of the legend in the plot. It can be
specified by keyword or in any way which is accepted by \code{xy.coords} (defalut
"topright")}

\item{Deggs_object}{an object of class Deggs generated from
\code{generate_subnetworks}}

\item{subgroup_variable}{column name in \code{metadata} which contains the
subgroup definition for each sample in \code{normalised_counts}}

\item{regression_method}{whether to use robust linear modeling for the
interaction p-values. Options are 'rlm' (default) or 'lm'}

\item{subgroups}{optional character vector indicating which subgroups
are used for comparison. If not specified, all subgroups in
\code{subgroup_variable} will be considered}
}
\value{
base graphics plot showing gene-gene regressions for each subgroup
while the pvalue of the interaction term of
\emph{gene A ~ gene B \} subgroup* is reported on top
}
\description{
Print the regression graph of a specific genes
}
