% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{print_regressions}
\alias{print_regressions}
\title{Plot regressions for a specific gene-gene interaction}
\usage{
print_regressions(deggs_object, gene_A, gene_B, legend_position = "topright")
}
\arguments{
\item{deggs_object}{an object of class \code{deggs} generated from
\code{generate_subnetworks}}

\item{gene_A}{character. Name of the first gene}

\item{gene_B}{character. Name of the second gene}

\item{legend_position}{position of the legend in the plot. It can be
specified by keyword or in any way which is accepted by \code{xy.coords} (defalut
"topright")}
}
\value{
base graphics plot showing gene-gene robust regressions for each
subgroup. The p value of the interaction term of
gene A ~ gene B \* subgroup is reported on top.
}
\description{
Plot regressions for a specific gene-gene interaction
}
\seealso{
\code{\linkS4class{deggs}}.
}
