% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{View_interactive_subnetwork}
\alias{View_interactive_subnetwork}
\title{Interactive shiny app to visualise subnetworks}
\usage{
View_interactive_subnetwork(deggs_object, subgroup, use_qvalues = TRUE)
}
\arguments{
\item{subgroup}{the subgroup of interest: the associated network will be
visualised}

\item{Deggs_object}{an object of class Deggs generated from
\code{generate_subnetworks}}
}
\value{
a network plot with selectable nodes and links
}
\description{
Explore subnetworks and interactively select regression plots
}
