% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGGs_printing_functions.R
\name{View_interactive_subnetwork}
\alias{View_interactive_subnetwork}
\title{Interactive shiny app to visualise subnetworks}
\usage{
View_interactive_subnetwork(deggs_object)
}
\arguments{
\item{deggs_object}{an object of class \code{deggs} generated from
\code{generate_subnetworks}}
}
\value{
a shiny interface showing networks with selectable nodes and links
}
\description{
Explore subnetworks and interactively select regression and box plots
}
